requirements.txt
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
decorator==4.4.2
Deprecated==1.2.14
distro==1.9.0
Flask==3.0.0
future==0.18.3
gcloud==0.18.3
googleapis-common-protos==1.61.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huepy==1.2.1
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
instabot==0.117.0
itsdangerous==2.1.2
Jinja2==3.1.3
jwcrypto==1.5.0
jws==0.1.3
MarkupSafe==2.1.3
mock==5.1.0
moviepy==1.0.3
numpy==1.26.2
oauth2client==4.1.2
openai==1.7.2
opencv-python==4.8.1.78
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
proglog==0.1.10
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
Pyrebase4==4.1.0
pyserial==3.5
PySocks==1.7.1
pytest==7.4.4
python-dotenv==1.0.0
python-jwt==2.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.29.0
requests-toolbelt==0.10.1
responses==0.24.1
rsa==4.9
schedule==1.2.1
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.18
Werkzeug==3.0.1
wrapt==1.16.0
